VanSight scripts

+++ VanSightHub/src/RelayController.cpp
@@ -1,19 +1,19 @@
 #include "RelayController.h"
 
-RelayController::RelayController(const int* pins, uint8_t count) 
+RelayController::RelayController(const int* pins, int count) 
     : _count(count) {
     // Allocate array of SimpleRelay pointers
     _relays = new SimpleRelay*[_count];
     
     // Create SimpleRelay objects for each pin
-    for (uint8_t i = 0; i < _count; i++) {
+    for (int i = 0; i < _count; i++) {
         _relays[i] = new SimpleRelay(pins[i]);
     }
 }
 
 RelayController::~RelayController() {
     // Clean up all relay objects
-    for (uint8_t i = 0; i < _count; i++) {
+    for (int i = 0; i < _count; i++) {
         delete _relays[i];
     }
     delete[] _relays;
@@ -24,45 +24,45 @@ void RelayController::begin() {
     // Nothing additional needed here
 }
 
-bool RelayController::isValidRelayNum(uint8_t relayNum) const {
+bool RelayController::isValidRelayNum(int relayNum) const {
     return (relayNum >= 1 && relayNum <= _count);
 }
 
-bool RelayController::turnOn(uint8_t relayNum) {
+bool RelayController::turnOn(int relayNum) {
     if (!isValidRelayNum(relayNum)) return false;
     
     _relays[relayNum - 1]->on();
     return true;
 }
 
-bool RelayController::turnOff(uint8_t relayNum) {
+bool RelayController::turnOff(int relayNum) {
     if (!isValidRelayNum(relayNum)) return false;
     
     _relays[relayNum - 1]->off();
     return true;
 }
 
-bool RelayController::toggle(uint8_t relayNum) {
+bool RelayController::toggle(int relayNum) {
     if (!isValidRelayNum(relayNum)) return false;
     
     _relays[relayNum - 1]->toggle();
     return true;
 }
 
-bool RelayController::getState(uint8_t relayNum) {
+bool RelayController::getState(int relayNum) {
     if (!isValidRelayNum(relayNum)) return false;
     
     return _relays[relayNum - 1]->isRelayOn();
 }
 
 void RelayController::allOn() {
-    for (uint8_t i = 0; i < _count; i++) {
+    for (int i = 0; i < _count; i++) {
         _relays[i]->on();
     }
 }
 
 void RelayController::allOff() {
-    for (uint8_t i = 0; i < _count; i++) {
+    for (int i = 0; i < _count; i++) {
         _relays[i]->off();
     }
 }
